+++ packages/svelte/src/internal/client/reactivity/deriveds.js
@@ -1,18 +1,9 @@
 /** @import { Derived, Effect } from '#client' */
 import { DEV } from 'esm-env';
-import {
-	CLEAN,
-	DERIVED,
-	DESTROYED,
-	DIRTY,
-	EFFECT_HAS_DERIVED,
-	MAYBE_DIRTY,
-	UNOWNED
-} from '../constants.js';
+import { CLEAN, DERIVED, DIRTY, EFFECT_HAS_DERIVED, MAYBE_DIRTY, UNOWNED } from '../constants.js';
 import {
 	active_reaction,
 	active_effect,
-	remove_reactions,
 	set_signal_status,
 	skip_reaction,
 	update_reaction,
@@ -125,7 +116,7 @@ function get_derived_parent_effect(derived) {
  * @param {Derived} derived
  * @returns {T}
  */
-export function execute_derived(derived) {
+function execute_derived(derived) {
 	var value;
 	var prev_active_effect = active_effect;